split code to subdirectories
