+++ react/src/lib/components/WellLogViewer/components/styles.scss
